Refactor: Move CLI modules to utils/cli and extract templates
- Moved all CLI modules from src/cli/ to src/utils/cli/ for better organization
- Extracted all templates (scripts, config, code, mcp) to external files in templates/
- Split easy-mcp-server.js into focused modules:
  * utils/cli/help.js - Help text display
  * utils/cli/templates.js - Template loading and processing
  * utils/cli/init-project.js - Project initialization logic
  * utils/cli/server-start.js - Server startup logic
  * utils/cli/utils.js - CLI utility functions
- Updated main easy-mcp-server.js to be clean 69-line entry point
- Updated package.json to include templates/ in npm package
- All templates now use placeholder replacement ({{PROJECT_NAME}})
- Template files are externalized for easier maintenance

Benefits:
- Better code organization and separation of concerns
- Templates can be edited without modifying source code
- Easier to maintain and understand
- All tests pass (18/18 init-command tests)
