+++ src/utils/cli/utils.js
@@ -1,0 +1,120 @@
+/**
+ * CLI Utility Functions
+ * Handles environment loading, hot reloading, and dependency installation
+ */
+
+const fs = require('fs');
+const path = require('path');
+
+/**
+ * Load .env files from user directory
+ * Loads files in order: .env.local, .env.development, .env
+ */
+function loadUserEnvFiles() {
+  const userCwd = process.cwd();
+  const envFiles = ['.env.local', '.env.development', '.env'];
+  
+  for (const envFile of envFiles) {
+    const envPath = path.join(userCwd, envFile);
+    if (fs.existsSync(envPath)) {
+      try {
+        require('dotenv').config({ path: envPath });
+        console.log(`📄 Loaded environment from ${envFile}`);
+      } catch (error) {
+        console.warn(`⚠️  Warning: Could not load ${envFile}: ${error.message}`);
+      }
+    }
+  }
+}
+
+/**
+ * Global env hot reloader instance
+ */
+let envHotReloader = null;
+
+/**
+ * Initialize environment variable hot reloader
+ * Watches .env files for changes and reloads them automatically
+ */
+function initializeEnvHotReloader() {
+  try {
+    const EnvHotReloader = require('../utils/loaders/env-hot-reloader');
+    envHotReloader = new EnvHotReloader({
+      debounceDelay: 1000,
+      onReload: () => {
+        console.log('🔄 Environment variables reloaded - some changes may require server restart');
+      },
+      logger: console
+    });
+    envHotReloader.startWatching();
+  } catch (error) {
+    console.warn(`⚠️  Could not initialize .env hot reload: ${error.message}`);
+  }
+}
+
+/**
+ * Auto-install missing dependencies
+ * Runs npm install if package.json exists
+ * @returns {Promise<void>}
+ */
+async function autoInstallDependencies() {
+  const userCwd = process.cwd();
+  const packageJsonPath = path.join(userCwd, 'package.json');
+  
+  if (!fs.existsSync(packageJsonPath)) {
+    console.log('📦 No package.json found - skipping auto-install');
+    return;
+  }
+  
+  console.log('📦 Checking for missing dependencies...');
+  
+  try {
+    const { spawn } = require('child_process');
+    
+    // Run npm install
+    const installProcess = spawn('npm', ['install'], {
+      cwd: userCwd,
+      stdio: 'inherit'
+    });
+    
+    return new Promise((resolve) => {
+      installProcess.on('close', (code) => {
+        if (code === 0) {
+          console.log('✅ Dependencies installed successfully');
+          resolve();
+        } else {
+          console.warn(`⚠️  npm install completed with code ${code}`);
+          resolve(); // Don't fail the server startup
+        }
+      });
+      
+      installProcess.on('error', (error) => {
+        console.warn(`⚠️  Could not run npm install: ${error.message}`);
+        resolve(); // Don't fail the server startup
+      });
+    });
+  } catch (error) {
+    console.warn(`⚠️  Error during auto-install: ${error.message}`);
+  }
+}
+
+/**
+ * Parse port configuration from environment variables
+ * @returns {Object} Configuration with port and mcpPort
+ */
+function parsePortArguments() {
+  const config = {
+    port: process.env.EASY_MCP_SERVER_PORT || 8887,
+    mcpPort: process.env.EASY_MCP_SERVER_MCP_PORT || 8888
+  };
+  
+  return config;
+}
+
+module.exports = {
+  loadUserEnvFiles,
+  initializeEnvHotReloader,
+  autoInstallDependencies,
+  parsePortArguments
+};
+
