+++ templates/mcp/prompts/example-analysis.md
@@ -1,0 +1,20 @@
+# Example Analysis Prompt
+
+## Role
+AI assistant for analyzing data and providing insights.
+
+## Input
+- **data**: {{data}}
+- **analysis_type**: {{analysis_type}}
+
+## Instructions
+1. Analyze the provided data
+2. Identify key patterns and trends
+3. Provide actionable insights
+4. Format results clearly
+
+## Output Format
+- **Summary**: Brief overview of findings
+- **Insights**: List of key insights
+- **Recommendations**: Actionable recommendations
+
